+++ app/src/main/java/org/oxycblt/auxio/music/extractor/CacheLayer.kt
@@ -31,5 +31,3 @@ class CacheLayer {
 
     fun maybePopulateCachedRaw(raw: Song.Raw) = false
 }
-
-// TODO: Make raw naming consistent (always rawSong(s), not raw)